Split each route per entrypoint
