+++ index.py
@@ -1,0 +1,35 @@
+# -*- coding: utf-8 -*-
+
+from scripts import tabledef
+from scripts import forms
+from scripts import helpers
+from flask import Flask, redirect, url_for, render_template, request, session
+import json
+import sys
+import os
+
+
+app = Flask(__name__)
+app.secret_key = os.environ['SECRET_KEY']
+
+@app.route('/', methods=['GET', 'POST'])
+def login():
+    if not session.get('logged_in'):
+        form = forms.LoginForm(request.form)
+        if request.method == 'POST':
+            username = request.form['username'].lower()
+            password = request.form['password']
+            if form.validate():
+                if helpers.credentials_valid(username, password):
+                    session['logged_in'] = True
+                    session['username'] = username
+                    return json.dumps({'status': 'Login successful'})
+                return json.dumps({'status': 'Invalid user/pass'})
+            return json.dumps({'status': 'Both fields required'})
+        return render_template('login.html', form=form)
+    user = helpers.get_user()
+    return render_template('home.html', user=user)
+
+
+if __name__ == "__main__":
+    app.run(debug=True, use_reloader=True)
