+++ settings.py
@@ -1,0 +1,32 @@
+# -*- coding: utf-8 -*-
+
+from scripts import tabledef
+from scripts import forms
+from scripts import helpers
+from flask import Flask, redirect, url_for, render_template, request, session
+import json
+import sys
+import os
+
+
+app = Flask(__name__)
+app.secret_key = os.environ['SECRET_KEY']
+
+
+@app.route('/settings', methods=['GET', 'POST'])
+def settings():
+    if session.get('logged_in'):
+        if request.method == 'POST':
+            password = request.form['password']
+            if password != "":
+                password = helpers.hash_password(password)
+            email = request.form['email']
+            helpers.change_user(password=password, email=email)
+            return json.dumps({'status': 'Saved'})
+        user = helpers.get_user()
+        return render_template('settings.html', user=user)
+    return redirect(url_for('login'))
+
+
+if __name__ == "__main__":
+    app.run(debug=True, use_reloader=True)
